+++ src/main/java/com/jvn/resume/Address.java
@@ -1,6 +1,5 @@
-package com.jvn;
+package com.jvn.resume;
 
-import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -9,10 +8,13 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 @Getter
 @Setter
 @NoArgsConstructor
-public class ToolSet {
+public class Address {
 
-  private String label;
-  private List<String> tools;
+  private String street;
+  private String additionalStreet;
+  private String city;
+  private String state;
+  private String zip;
 
   @Override
   public String toString() {
