+++ src/main/java/com/jvn/resume/DateRange.java
@@ -1,6 +1,5 @@
-package com.jvn;
+package com.jvn.resume;
 
-import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -9,18 +8,15 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 @Getter
 @Setter
 @NoArgsConstructor
-public class Experience {
+public class DateRange {
 
-  private String title;
-  private DateRange tenure;
-  private String company;
-  private Address address;
-  private List<String> technologyUsed;
-  private List<String> highlights;
+  private SimpleDate startDate;
+  private SimpleDate endDate;
 
   @Override
   public String toString() {
     return ToStringBuilder.reflectionToString(this);
+
   }
 
 }
