+++ src/main/java/com/jvn/resume/Experience.java
@@ -1,6 +1,6 @@
-package com.jvn;
+package com.jvn.resume;
 
-import java.util.Objects;
+import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -9,15 +9,18 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 @Getter
 @Setter
 @NoArgsConstructor
-public class DateRange {
+public class Experience {
 
-  private SimpleDate startDate;
-  private SimpleDate endDate;
+  private String title;
+  private DateRange tenure;
+  private String company;
+  private Address address;
+  private List<String> technologyUsed;
+  private List<String> highlights;
 
   @Override
   public String toString() {
     return ToStringBuilder.reflectionToString(this);
-
   }
 
 }
