+++ src/main/java/com/jvn/resume/ToolSet.java
@@ -1,5 +1,6 @@
-package com.jvn;
+package com.jvn.resume;
 
+import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -8,14 +9,10 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 @Getter
 @Setter
 @NoArgsConstructor
-public class Personal {
+public class ToolSet {
 
-  private String firstName;
-  private String lastName;
-  private Address address;
-  private String phoneNumber;
-  private String email;
-  private String website;
+  private String label;
+  private List<String> tools;
 
   @Override
   public String toString() {
